requirements.txt
azure-core==1.38.0
azure-cosmos==4.14.6
azure-functions==1.24.0